requirements.txt

fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
google-adk>=0.1.0
pyyaml>=6.0
